Move openModal to the global actions
